requirements.txt
alembic==1.18.4
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
charset-normalizer==3.4.7
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
tabulate==0.10.0
typing_extensions==4.15.0
urllib3==2.6.3
